get_env(name) mismatches FOOBAR when searching FOO
When searching for environment variable FOO, get_env(name) will declare a match
upon finding name FOOBAR.

Also, if an environment variable name is matched, but its value is a string of
spaces (e.g. "   "), then "out-of-range" will be thrown.

Also, why is the logic trying to trim spaces in the first place?  Spaces are
legal within an environment variable.

Also, the get_env() function uses both env_var.length() and env_var.size() which
do exactly the same thing -- it should use one or the other for consistency.